Implement NetworkBuilder for specifying a network
The shape of a network (i.e. the number of layers and the number of
units in each layer) can now be specified using `NetworkBuilder`. The
file `examples/create_network.rs` contains an example of how this API is
used.

The notable thing about this design is that the number of units in each
layer is encoded into its type, and the number of units in the input and
output layers of the network are in turn encoded into the `Network`
type. Theoretically this design should allow for more flexibility later,
such as by chaining networks together or using entire networks as though
they are single units or groups of units within one layer.

By structuring a network explicitly as a list (a linked list, in fact)
of layers, I've made a deliberate decision to prioritise several
different things over performance. In particular, this design should
allow for quite extensive visualisation of the structure of, and flow of
data through, networks farther down the line.

In order to do the above-mentioned encoding of the size of each layer
into its type, the version of Rust needed to build the crate has been
pinned to a nightly version. Thankfully, this design only needs to
generalize over `usize` types, so it needs only the already-stable
`min_const_generics` feature, and not the incomplete `const_generics`.
Hopefully soon we'll be able to move back to a stable version.